refactor(core): consolidate recents methods into AssetService
