+++ packages/core/src/index.ts
@@ -28,4 +28,3 @@ export * from './gotenberg/gotenberg'
 export * from './mcp/mcp-service'
 export * from './quota/quota-service'
 export * from './kanban/kanban'
-export * from './recents/recents'
